Analytical French & Wilson scaling
Based on the method of Read & McCoy (2016) Acta Cryst D72(3): 375-387. 
Implemented in Python for the time being since I haven't found an easy, 
compact standalone C++ implementation of erfc and parabolic cylinder 
functions (these are available in Boost, C++17 and the Intel MKL, so 
porting will be easily possible if necessary in future). However, since 
this scaling is something that typically only needs to be done once the 
Python code will probably be fast enough.