Auto choose free flags and reflections when called from command line
ChimeraX convention is that command-line calls do not launch GUI 
windows. When loading .cif or .mtz files with multiple possible free 
flag arrays and/or multiple reflection data arrays, these will now be 
automatically chosen by default (with warnings sent to the log). For 
free flags, the choice will be the first flag array with 'free' in the 
lowercase-cast name, or the first flag array otherwise. For experimental 
reflections, the order of precedence is intensities > amplitudes > 
anomalous intensities > anomalous amplitudes.