Fcalc maps hidden from default
While there is no foolproof way of programmatically determining whether 
a given F/phi dataset is an FC map, we can hopefully rely on most 
writing programs being relatively sensible about it. The most common 
names for the F column (after casting to uppercase) are "FC", "F_CALC" 
or "FCALC", so we'll work under the assumption that anything starting 
with "FC" or containing "CALC" is an FC map.