Fixed bug causing malformed symmetry bonds in some cases
When finding symmetry atoms, it is not uncommon for a symmetry operator 
to appear as a candidate but turn out to have no atoms associated with 
it. The code was skipping forward by one atom for each "empty" symmetry 
operators, causing some bonds to be drawn between incorrect symmetry 
partners and others to not be drawn at all.