Ignore symmetry info if mmCIF file has no 'exptl' table
The cell and space group information in a cryo-EM (or any other 
non-crystallographic) model is not to be trusted. The problem is, not 
every mmcif file *tells* us of its provenance. If it doesn't explicitly 
state that it's a crystallography model, it's best to ignore its 
symmetry information and instead get it from the structure factors when 
and if they're loaded.