Improve stability of anisotropic gaussian scaling function
The ResolutionFn and ResolutionFn_nonlinear classes use unrestrained 
Newton-Raphson minimisation, which knows nothing about the concept of 
positive definiteness. This can cause the anisotropic Gaussian scaling 
function to head into quite wild territory in some cases. Solution (for 
now) is to first fit the isotropic Gaussian, and use the parameters from 
that to set the initial guesses for the anisotropic Gaussian parameters.