Parallelised fetching of map data from Xmap
Fetching a big box on a single thread is quite slow (10s of 
milliseconds) due to all the necessary symmetry lookups. With some 
fairly heavy refactoring it should be possible to improve things 
significantly by fetching the data for all isomorphous maps together, so 
the time-consuming step of converting the Map_Reference_Coord to an 
index only has to be done once for each grid point.